refactor(ui/waveform_view): split interactions modules and extract selection drag handling
