+++ src/egui_app/ui/waveform_view/destructive_prompt.rs
@@ -106,4 +106,3 @@ impl EguiApp {
         });
     }
 }
-
