+++ src/egui_app/ui/waveform_view/interactions/mod.rs
@@ -1,0 +1,7 @@
+mod pointer;
+mod scroll_zoom;
+mod scrollbar;
+
+pub(super) use pointer::handle_waveform_pointer_interactions;
+pub(super) use scroll_zoom::handle_waveform_interactions;
+pub(super) use scrollbar::render_waveform_scrollbar;
