+++ src/egui_app/ui/waveform_view/overlays.rs
@@ -64,4 +64,3 @@ pub(super) fn render_overlays(
         );
     }
 }
-
