+++ src/egui_app/ui/waveform_view/selection_geometry.rs
@@ -73,4 +73,3 @@ mod tests {
         assert!(end_edge_rect.bottom() <= handle_rect.top());
     }
 }
-
